fix: split field invariant code for readability ratchet
Move field invariant check into field_constraints.rs, extract tests to
field_invariant_tests.rs, compact constraint constructors, and update
the readability baseline.
