+++ .ci/readability-baseline.env
@@ -1,7 +1,7 @@
 # Generated by scripts/readability-ratchet.sh
-PROD_RS_TOTAL=307
-PROD_FILES_GT300=98
-PROD_FILES_GT500=47
+PROD_RS_TOTAL=310
+PROD_FILES_GT300=100
+PROD_FILES_GT500=48
 PROD_FILES_GT1000=0
 PROD_MAX_FILE_LINES=987
 PROD_MAX_FILE_PATH=crates/temper-spec/src/automaton/toml_parser.rs
